requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.3.2
flask-restx==1.1.0
importlib-metadata==6.7.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
pkgutil_resolve_name==1.3.10
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.0
rpds-py==0.9.2
Werkzeug==2.3.6
zipp==3.15.0
